+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -305,7 +305,7 @@ internal static class Instrumentation
                     DelayedInitialization.Traces.AddGrpcClient(lazyInstrumentationLoader, pluginManager);
                     break;
                 case TracerInstrumentation.SqlClient:
-                    DelayedInitialization.Traces.AddSqlClient(lazyInstrumentationLoader, pluginManager);
+                    DelayedInitialization.Traces.AddSqlClient(lazyInstrumentationLoader, pluginManager, tracerSettings);
                     break;
                 case TracerInstrumentation.Quartz:
                     DelayedInitialization.Traces.AddQuartz(lazyInstrumentationLoader, pluginManager);
@@ -320,7 +320,7 @@ internal static class Instrumentation
                 case TracerInstrumentation.MySqlData:
                     break;
                 case TracerInstrumentation.EntityFrameworkCore:
-                    DelayedInitialization.Traces.AddEntityFrameworkCore(LazyInstrumentationLoader, pluginManager);
+                    DelayedInitialization.Traces.AddEntityFrameworkCore(LazyInstrumentationLoader, pluginManager, tracerSettings);
                     break;
                 case TracerInstrumentation.StackExchangeRedis:
                     break;