+++ src/OpenTelemetry.AutoInstrumentation/Loading/Initializers/EntityFrameworkCoreInitializer.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #if NET6_0_OR_GREATER
+using OpenTelemetry.AutoInstrumentation.Configurations;
 using OpenTelemetry.AutoInstrumentation.Plugins;
 
 namespace OpenTelemetry.AutoInstrumentation.Loading.Initializers;
@@ -9,18 +10,23 @@ namespace OpenTelemetry.AutoInstrumentation.Loading.Initializers;
 internal class EntityFrameworkCoreInitializer : InstrumentationInitializer
 {
     private readonly PluginManager _pluginManager;
+    private readonly TracerSettings _tracerSettings;
 
-    public EntityFrameworkCoreInitializer(PluginManager pluginManager)
+    public EntityFrameworkCoreInitializer(PluginManager pluginManager, TracerSettings tracerSettings)
         : base("Microsoft.EntityFrameworkCore")
     {
         _pluginManager = pluginManager;
+        _tracerSettings = tracerSettings;
     }
 
     public override void Initialize(ILifespanManager lifespanManager)
     {
         var instrumentationType = Type.GetType("OpenTelemetry.Instrumentation.EntityFrameworkCore.EntityFrameworkInstrumentation, OpenTelemetry.Instrumentation.EntityFrameworkCore")!;
 
-        var options = new OpenTelemetry.Instrumentation.EntityFrameworkCore.EntityFrameworkInstrumentationOptions();
+        var options = new OpenTelemetry.Instrumentation.EntityFrameworkCore.EntityFrameworkInstrumentationOptions
+        {
+            SetDbStatementForText = _tracerSettings.InstrumentationOptions.EntityFrameworkCoreLSetDbStatementForTest
+        };
 
         _pluginManager.ConfigureTracesOptions(options);